feat(time): Allow user to specify time format
This commit adds logic so that the user can specify their own code for how they want the time to be formatted. The most obvious options are 12 hour and 24 hours, but the user will have complete flexability for how they want to format it. Up to and including setting it to an empty string